pyproject.toml
[project]
name = "mcp-sqlify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.5.1",
    "langchain-openai>=0.3.16",
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
